requirements.txt
Cython==0.29.12
pymssql==2.1.4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
